pyproject.toml
[tool.poetry]
name = "craes_pyclient"
version = "0.1.0"
description = "A static CLI client for CRAES"
authors = ["Daniel Bischof <daniel.bischof@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
websockets = "^8.1"
pycapnp-async = "^0.7.0"
cython = "^0.29.14"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
python-language-server = "^0.31.2"
pylint = "^2.4"

[tool.poetry.scripts]
client = "craes_pyclient.main:run_client"
testserver = "craes_pyclient.test_server:run_sink_server"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
